handles, 6 months, today line & scrubber, comments
Handles are now translucent circles instead of ellipses.

6 months now accurately only displays 6 months.

The scrubber’s initial width is now made so that it lines up with today
(note: this is hard coded in… update if you ever change the today
date). This means that the reset button now works again.

Various comments in code